Refactor: Move mob functions to a separate file
Moved mob-related functions from js/main.js to a new js/mobs.js file.
Updated index.html to include the new script.
This change improves code organization without altering functionality.
