+++ index.html
@@ -265,5 +265,7 @@
     <script src="js/audio-player.js"></script>
     <script src="js/video-player.js"></script>
     <script src="js/worker.js"></script>
-    <script src="js/main.js"></script></body>
+    <script src="js/mobs.js"></script>
+    <script src="js/main.js"></script>
+</body>
 </html>
